refactor: break up useChatService.ts
